+++ app/Views/layouts/partials/footer.php
@@ -159,5 +159,5 @@ $siteJsVersion = is_file($siteJsPath)
     ? (string) (md5_file($siteJsPath) ?: filemtime($siteJsPath))
     : (string) time();
 ?>
-<script src="<?= base_url('assets/js/alpine.min.js') ?>" defer></script>
-<script src="<?= base_url('assets/js/site.js?v=' . $siteJsVersion) ?>" defer></script>
+<script <?= csp_script_nonce() ?> src="<?= base_url('assets/js/alpine.min.js') ?>" defer></script>
+<script <?= csp_script_nonce() ?> src="<?= base_url('assets/js/site.js?v=' . $siteJsVersion) ?>" defer></script>